Make `lora_alpha` a float (#10004)
There's nothing fundamentally integer or continuous about `lora_alpha`. While artificially
restricting to integers may not have a huge impact in typical (human) usage, needlessly discretized
parameters can be tricky for hyperparameter search using cleverer-than-grid-search methods
(e.g. optuna, Ax).

Now, the `peft` dependency admittedly does have type hints expecting `int`. It seems to deal with
floats fine. `peft` upstream documents in their CONTRIBUTING.md a dislike for typing-only fixes.
I have nevertheless raised this point, offering to create a PR that changes the types, here:
https://github.com/huggingface/peft/issues/3615

+++ swift/megatron/arguments/megatron_args.py
@@ -463,7 +463,7 @@ class MegatronTunerMixin:
 
     # lora
     lora_rank: int = 8
-    lora_alpha: int = 32
+    lora_alpha: float = 32.0
     lora_dropout: float = 0.05
     lora_bias: Literal['none', 'all'] = 'none'
     lora_dtype: Literal['float16', 'bfloat16', 'float32', None] = None
